feat: changing pid waiting
