+++ v2/Makefile
@@ -6,7 +6,7 @@
 #    By: maalexan <maalexan@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/01/22 11:13:39 by maalexan          #+#    #+#              #
-#    Updated: 2023/09/02 16:12:50 by maalexan         ###   ########.fr        #
+#    Updated: 2023/09/05 12:23:31 by maalexan         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -47,6 +47,7 @@ FUN			:=	main.c \
 				env_utils.c \
 				exec_finder.c \
 				exec_run.c \
+				executor.c \
 				input_utils.c \
 				input_validator.c \
 				parser.c \
